caos.backends.pgsql: Always use raw prepared statements except for scrolling cursors
The default behaviour of py-postgresql is to use a Cursor for every
query that is run outside a transaction.  While the reasoning for this
is not completely available, it is fairly straigtforward to PREPARE the
statements directly and use statement_from_id(), unless there is a need
for an actual cursor.