caos: Fix several issues in select() pathspec mechanism
Sub-pathspec for recursive specs is now handled correctly.  Previously,
entries in sub-pathspec could have been omitted from the query due to
buggy recursion control.

Pathspecs are now also normalized correctly so that there are no
duplicate entries in the list for each pointer.  If multiple entries
are found, they are merged and a common target is selected using the
_least generic_ inheritance path, i.e. a descendant target class would
win.  This is in contrast to the new behaviour of Source.follow() that
would merge multiple targets using the _most generic_ path.  The
difference is that pathspecs are usually written to _narrow_ the scope
of the query.