caos: Turn Prototype.base into a list of references instead of a list of names
Prototype.base was the only prototype field that was still referencing
other prototypes by name, thus, creating considerable difficulties when
the list of parent prototypes needed to be resolved and the schema
wasn't available.  Most of the machinery to handle references in
prototype fields is already in place, so this reduced mostly to fixing
sites of reading of and writing to the field, which there was quite a
number of.