utils.lang.import_.loader: get_module_version() returns 0 if module file no longer exists
It is possible in a reload scenario that a loaded module file no longer
exists.  In order for the get_modified/reload machinery to work
correctly, the modversion for such modules is set to 0, signaling the
need to reload the module and its dependencies.  A failure to reload due
to missing file can then be handled nicely by the reload function.