caos: Stop caching inheritance children in prototypes
Currently to support descending inheritance resolution, prototype
children are cached in the _children property directly in the parent
prototype.  This violates the protomodule immutability rule, whereby
a fully-loaded module contents must not be altered by other modules.
Thus, move the inheritance tracking to schema level.  The only exception
to this is virtual concepts, which do have to have their children
listed, as they are not recorded into bases, and do not break the module
immutability rule.

As a consequence, Prototype.children() will now require protoschema
as an argument.  Additionally, the `recursive' argument is gone, and
Prototype.descendants() method should be used instead.