caos: Stop creating tables for virtual concepts at runtime
Initial reasoning behind runtime creation of tables for virtual concepts
was to offload most query logic to native database table inheritance
mechanisms.  Als, while theoretically sound, in practice this turned out
to be a huge hassle maintenance-wise.  Moreover, unpredictable DDL in
runtime session is causing postgres protocol exceptions quite often.

Fortunately, query plans for equivalent UNION queries are generated
properly in Postgres 9.2, so virtual concepts can be represented by
inline UNION subqueries.  The only bit of functionality lost is the
.update() of a virtual concept selector, which, AFAIK, isn't currently
used in any of the codebases, and can be replaced with a set of UPDATE
queries for each non-virtual child.