utils.lang.python: Improved performance of stack size calculation.
No longer using a recursive method to calculate the maximum stack size
for a piece of code. This has the advantage of handling larger chunks of
code without running out of memory. It should also be somewhat faster as
list.pop() is probably more efficient than a recursive function call.