utils.lang.python: Handling loops when calculating max stack size.
We don't need to explicitly "unsee" any ops because we keep a copy of
this data every time we make a jump. We only need to "unsee" ops when we
consider an alternate computation path for a conditional jump. This case
is handled by restoring the pre-jump state of seen ops.