caos: Distinguish between computable and access control rewrites
The computable and access control rewrites need to be distinguished as
they have different application semantics: filter rewrites need not be
applied on derived paths (e.g. in subqueries), as a corresponding parent
path would already satisfy filter's requirement.

On the other hand, computable rewrites must be applied to every
reference in the tree regardless of its position.  It is up to the
backend transformer to apply optimizations by grouping computable
expressions.