utils.lang.js: Add two new datatypes: sx.ds.Map and sx.ds.Set
Those two are almost drop-in replacements of proposed JS Harmony Map &
Set objects, with just a few exceptions:

- No difference between +0 and -0
- No 'delete' method, instead use 'del'.  That's because old browsers
  won't support keywords as method names.

The implementation uses native Map and Set if they are available, but so
far this is true only for Firefox.

And a word of caution: don't use theese until absolutely necessary, the
performance is going to be awful.