utils.lang: Implement module tagging system
It is now possible to set an arbitrary number of tags on any module.
The motivation is to create an ability to differentiate and group
loaded modules.

The tags is normally a collection of ModuleTag subclasses in
__mm_module_tags__ module attribute.  The tags may either be set
directly, or via a tagging policy specified in an ancestor package.
The latter mechanism is most powerful and supports matching of module
names with glob patterns.  The most straightforward way to specify a
package tagging policy is to have an __init__.yml with
utils.lang.package.schema.Package schema::

     %schema metamagic.utils.lang.package.schema.Package
     %from ..tags import TestTag, TestTag2, TestTag3
     ---

     tags:
         '**.ham': []
         'foopkg.py': [TestTag3]
         'foopkg.*.*.test': [TestTag]