utils.lang: Allow Runtimes to be registered for multiple Languages
It is now possible to pass a tuple to the "languages" class argument
(renamed from language).  "register" class argument has been renamed to
"abstract" with the default reversed.

Additionally, make it possible for the Runtime to refuse to create any
derivatives by raising NotImplementedError from new_derivative() which
is now the default behaviour.