caos.frontends.javascript: Handle derived virtuals on far end of inbound links
It is possible that a query path contains a dynamically derived virtual
node on the far end of an inbound link.  Since the original reflected
schema does not contain the virtual, we need to supply the necessary
information in query metadata so that the link is resolved and unpacked
properly on the client.