caos: Add implementation of re:match() function
re:match(regexp, string[, flags) returns a tuple of values for each
matching group in `regexp`, if `regexp` matches `string`, otherwise it
returns None.  If no capturing groups are defined in `regexp` the return
value would be a tuple containing a single string value equal to the
entire matching string.