caos: Implement generic prototype attributes
Prototype Attributes is now an official way to add extra metadata to
Caos schema objects.  An attribute is a typed and named value.
Two new object types are introduced to the schema: attributes and
attribute values.  The former define attributes, the latter specify
actual attribute values and attach to the relevant schema object.
Attributes are typed with an option to define constraints on values.

builtins.decimal and builtins.sequence have been converted to use
attributes for precision and rounding, and sequence pattern
correspondingly.