caos: Reimplement computables as a Pointer class trait
The current implementation of computables is very broken.  It is a
separate Caos system class, yet it attempts to masquerade as different
Pointer classes introducing a lot of unnecessary complexity.

"Computable" is, really, just another pointer trait.  Thus, reimplement
computables so that any read-only pointer with a default value is trated
as a computable value.

While at it, switch the previous ad-hoc computable query rewrite path to
a common query rewrite engine.

This commit also removes more lines than it adds, which is always a good
thing.