caos: New, deterministic specialized pointer name generation
Current scheme (non-qname!hash(source, target)!canonical:name) has
several major shortcomings: a) it is not reversible, b) both source and
target have to be known at the time of name derivation, which may not be
possible.

To fix this, switch to a new generic scheme:

   non-qname@canonical:name@suffix

where "non-qname" is a non-qualified name of an attribute,
"canonical:name" is attribute's fully-qualified canonical name with dots
replaced by colons, and "suffix" is a quoted name of attribute's
subject, where colons are doubled and dots are replaced with colons.

This scheme allows consistent encoding of arbitrary level of nesting of
object attributes, i.e. an attribute value on a link property of a
specialized link.