caos.backends.pgsql: Fix a long-standing issue with dangling joins
Currently, a weak path, when translated to a join chain, can produce
extraneous rows due to a dandgling join on the target side of link
mapping.  This can be caused by either partial data (actual dangling
link), or by query constraints, such as link target type restriction.
The latter case is the most common, and most troublesome, as it requires
manual addition of NOT IS NONE checks to produce correct result.

This changes the behaviour of transformer so that it always uses INNER
JOINs on the target side of link join, and also groups target joins
together, eliminating extraneous dangling rows.