utils.fs: Make it possible to resolve a file path by id only
Current fs implementation requires both the id and the name of a file to
obtain its path.  This may be impractical in situations where only the
id is known, so allow buckets to declare that the name is optional when
resolving the file path.  This means that the id portion of each file is
assumed to be unique, so that a symbolic link can be safely created to
actual file path.