caos.backends.yaml: Raise an appropriate error on invalid fixture data
Fixture schema allows using YAML references to refer to targets defined
in the same fixture earlier.  However, the nature of YAML constructor
makes this case indistinguishable from an incorrect definition of an
entity without a concept class.  For example:

Class1: &E1
    name: Spam

Class2:
    link: *E1

Class3:
    link:
       name: Spam

For the constructor, the second and the third cases are equivalent,
however, only the first of the two is correct, as it refers to an
entity that the fixture loader knows the concept class for.  Teach
the fixture loader to detect the latter case and raise an appropriate
error.