caos: Switch index handling to computable/constraint expression machinery
Historically, index specifications were the first kind of CaosQL
expressions used in schemas.  With subsequent introduction of
computables and constraints, the schema expression machinery was
generalized and upgraded, but index expressions were never ported.
This commit fixes that.

A consequence of this change is that index expressions must now include
explicit anchor "self" when referring to pointers of the host
source.  For example:

        indexes:
            - my_property

becomes:

        indexes:
            - self.my_property

This is similar to computables and constraints.