caos: Generalize framework for schema policy
A policy is a generic trait that attaches to a schema object and
dictates its behaviour in a defined situation.  There may be two types
of policies: internal policy, which is defined directly on a schema
object, and an external policy, which may be defined separately from a
schema object itself.

This commit only deals with reorganization of policy schema and adds
support for internal policies.  Currently, only links have support for
policies.

The former notion of CascadePolicy has been removed and replaced by this
new framework.