caos: Introduce support for CaosDDL in deltas
This brings in a major change in how schema changes are represented.

All transitions in schema states, including initial creation, is stored
and interpreted internally as an ordered tree of "delta commands" which
are replayed by the backend to move the schema from state A to state B.

Currently, the delta command tree is simply dumped directly into a
migration file.  This approach has a number of drawbacks:

  * migration file size inefficiency;
  * poor migration file readability;
  * exposure of core internals in migration files, including
    references to implementation classes.

The last point is especially painful as it makes any core refactoring an
unnecessary backward-compabibility-forward-delta-migration exersise.

In order to fix these issues we introduce support for CaosDDL: a
CaosQL syntax for manipulation of schema objects.  Like CaosQL, it
follows the spirit of the corresponding SQL part by introducing
CREATE/ALTER/DROP statements for each type of schema object.

Example:

    USING
        builtins := NAMESPACE metamagic.caos.builtins,
        extras := NAMESPACE metamagic.caos.extras
    CREATE ABSTRACT CONCEPT [extras.Dictionary]
           INHERITING ([extras.NamedObject], [extras.BaseDictionary])
    {
        SET is_virtual = False
        CREATE REQUIRED LINK [extras.name] TO [builtins.str] {
            SET mapping = '11'
            SET readonly = False
            SET search_weight = 'A'
            CREATE ABSTRACT CONSTRAINT [builtins.unique] {
                SET errmessage = 'invalid {subject}'
                SET expr := (not EXISTS (subject))
                SET finalexpr := (not EXISTS (subject))
                SET localfinalexpr := (not EXISTS (subject))
            }
        }
    }

Delta commands are still used internally and are converted transparently
to/from CaosDDL statements.

This new foundation also opens the door to interesting new capabilities,
such as interactive schema manipulation, which will be the focus of
future work.