caos.caosql: New syntax for namespace aliases and path aliases
This changes the syntax for specifying path aliases in CaosQL.
Currently, an alias for a path expression can be specified in curly
braces directly after the path.  For example:

    SELECT
        P.id
    WHERE
        Some.path{P}.value = 1

The readability of this syntax is subpar. The alias is both in the
middle of a potentially long path expression, and can also be specified
in a section of the query that is lexically below the section that is
using the alias.  In the example above the alias is declared in WHERE
clause and is used in select target list.

To address these issues and to improve readability, the alias
declaration syntax is changed to the following:

    USING
        P := Some.path
    SELECT
        P.id
    WHERE
        P.value = 1

Note that this new syntax shares the USING section with the
declarations of namespace aliases, which has also been modified to match
expression alias declarations better:

    USING
        NAMESPACE my.namespace,
	Alias := NAMESPACE my.namespace2
    ...