caos: Stop persisting virtual concepts
Whenever a link is declared to have multiple targets, an abstract
parent concept is automatically created to represent those multiple
targets.  Since the concept has not been declared explicitly, it is
called a _virtual_ concept.  Currently, such concepts are persisted in
the schema, complicating things.

Since the virtual concept can be re-created on-demand, stop persisting
it, and simply re-create it when loading the schema.