Rip out most of JavaScript infrastructure
The JS adapter code has bitrotted for a while and is mostly outdated
feature-wise.  Remove all adapters and the JS code they depend on, as
everything has a better-quality JPlus adapter.

The JS import support is retained, as is the 'sx' helper library, as it
is still used in a few places that don't have an adequate JPlus
equivalent yet.