caos: Refactor schema implementation, improve backend/frontend separation
The implementation for the core schema objects (proto.py), and their
mutations (delta.py) has grown rather large over time.  Since the DDL
introduction removed the exposure of schema organization on app code,
it's now safe to reorganize the code for better clarity and modularity.

Additionally, this refactoring moves things around to draw a clearer
separation line between the backend and the frontend, with only a few
bits of frontend-specific code to be dealt with.