Allow any Atom as an Attribute type
Previously, only a small set of atom-like types were allowed to be used
as attribute types.  Additionally, attributes had their own
implementation for constraints.  The reason for this is largely
historical, and it is now possible to simply allow any Atom as an
Attribute type.  This simplifies attribute implementation considerably.