Initial implementation of the rewritten query compiler
This is the initial effort to replace the current query compiler with the
new, more maintainable, efficient implementation.

The new compiler uses a significantly simpler IR.  Unlike the current
implementation, the new IR is not influenced by any aspects of SQL
translation and specifics of data storage, as those concerns can
be handled much better by the SQL backend.

This is just an initial implementation capable of translating simple
queries into unoptimized SQL.  The strategy is to keep the original
implementation around until the new compiler reaches complete feature
parity.