compiler: Reimplement PostgreSQL AST and code generation
As the final compiler rewrite step there is now completely new
PostgreSQL AST and codegen which have been made much closer to
the native PostgreSQL parse node structure.  The new compiler
is now targeting this new codegen backend.

This also flips the switch to use the new compiler in CLI and in tests
by default.  xfail was also flipped on a few tests fixed or broken by
the switch.  The majority of new xfails are around the handling of
the RETURNING clause in DML statements.