compiler: Rewrite path scoping.
The current implementation attempts to maintain a list of subqueries for
every relation and then injects path scope conditions itno those subqueries
once the parent query has been fully formed.

This approach is problematic, as keeping a correct track of subqueries
is non-trivial, especially when views are involved.

The new implementation in this commit takes a different apporach: every
query defines its path scope _before_ any of the subqueries are formed,
and subqueries simply take the scope from the context as needed and at
the time of their creation.

Note that this change has introduced a regression into the handling
of references to constant expressions in views, but that's largely due
to the removal of scope hacks around view handling.  The proper fix for
these cases will follow in a separate commit.