compiler: Implement expression result cardinality check framework.
In certain contexts EdgeQL does not allow an expression to return
non-singleton sets.  Such cases need to be checked by the query
compiler.

This commit implements the framework to infer expression cardinality
based on the current context.  Checks have been implemented to enforce
singleton sets in ORDER BY, LIMIT and OFFSET clauses.  Appearance of
path expressions in the GROUP BY clause implies that the result of the
path expression should be considered a singleton in the enclosing query.