compiler: Fix cardinality of paths referring to constant expressions in views.
We need to make sure that the target expression is computed at least N
times, where N is the cardinality of the path source set.  However,
we cannot simply join the source set as it might have already been
injected via the parent range scope mechanism.  To determine whether the
source join is necessary, do a deep search on the target expression.