edgeql: Change 'foo AS MODULE ...' to 'foo := MODULE ...'
The fundamental reason for this change is simplicty in terms of mental
effort required to learn EdgeQL. Having `AS` and `:=` used in different
contexts increases the burden on the user to remember the correct syntax
in any particular case. We also do not have a consistent rule for using
one vs. the other.

Options that were considered and discarded:

- Use some notion of imperative vs. declarative semantics to tell `:=`
  and `AS` apart. (EdgeQL is fundamentally declarative, so it makes
  little sense.)

- Use a notion of pure lexical macro-like aliasing vs. sub-queries and
  structs. (Problem is that aliasing makes it non-trivial to understand
  scoping and therefore semantics of a statement. So it makes no sense
  to insist that pure lexical aliasing is somehow not interacting with
  semantics.)

- Arbitrarily define structs, kwargs and PL/EQL as using `:=`, whereas
  all other cases use `AS` since it's more declarative. (The distinction
  is too arbitrary, and it places a greater burden on the user to
  remember the correct usage.)