edgeql: Rework support for grouping.
Executive summary: SELECT ... GROUP BY is now GROUP ... BY ... RETURNING

The main motivation for this change is the simplification of the mental
model for EdgeQL queries.  The new doctrine dictates that each statement
declares certain linear transformation of data where the data flows in
the order of statement clauses: top to bottom.  Specifically, for GROUP
statements:

    GROUP
        <Expr>                   -  declare a set of values (essentially
                                    a View)
    BY
        <Expr>[,...<Expr>]       -  partition the set into buckets

    RETURNING
        ...                      -  compute a new set of values using
                                    aggregation on grouped buckets
    WHERE
        ...                      -  filter the aggregated set

    ORDER BY
        ...                      -  sort the filtered set

    OFFSET/LIMIT
        ...                      -  slice the sorted set

In order to support this flow, expressions in output clauses can now
have an optional alias:

    GROUP
        alias := {
            foo := 'bar'
        }
    BY
        alias.foo
    RETURNING
        output := {
            foo_list := array_agg(alias.foo)
	}
    ORDER BY
        len(output.foo_list)

The output expressions essentially appear as views for the clauses that
follow them.

This commit only adds support for this behaviour to the input GROUP
expression, support for SELECT will be in a follow up commit.

[ Initial syntax work by Victor. ]