edgeql: Change statement syntax.
The change takes the idea used for GROUP statement and generalizes it to
all other statements. Each statement declares certain linear
transformation of data where the data flows in the order of statement
clauses: top to bottom. The doctrine now includes DELETE, UPDATE and
INSERT statements. In accordance with this mental model the keyword
WHERE has been renamed to FILTER to better reflect the change and to
highlight the difference from how SQL uses the WHERE clause.

Another syntactic change that reinforces the notion of top-to-bottom
data flow is that all statements can use aliases for their subject and
RETURNING clauses. This makes it possible for other clauses to refer to
data they are intended to manipulate in a natural way.