Major compiler rework fixing many remaining issues violating the model.
The main theme of this rework is the overhaul of path var routing in the
compiler as well as the general path scope handling approach, which
resulted in a much more robust compilation pipeline.

This allows fixing many cases where EMPTY was not handled appropriately,
most notably in boolean logic expressions, where an EMPTY operand will
now always cause EMPTY result.  This is notably different from SQL's
behaviour, but is consistent with EdgeQL set theory.