Remove support for RETURNING clause from DML statements.
The INSERT, UPDATE, and DELETE statements no longer support the
RETURNING clause.  Instead, a wrapping SELECT statement should be used
with the DML statement as either a subquery or a view, so

    INSERT Object RETURNING Object {name};

becomes

    SELECT (INSERT Object) {name};

or

    WITH I := (INSERT Object) SELECT I {name};

The rationale is that the RETURNING clause does not provide anything
that a wrapping SELECT cannot do.

Additionally, bare DML statements now return the number of records
affected.