compiler: Clean up the handling of tuples.
There is now only one tuple type internally, and, semantically, bare
tuples behave the same way as named tuples.  Essentially, an bare
tuple is a named tuple where element names are natural numbers starting
at 0, which indicate the position of the element within the tuple.

A follow-up commit will add syntactic support for referring to the
elements of bare tuples.

The <Expr> IN (<Expr>,...) syntax has been removed, as the RHS is parsed
as a tuple, and the IN is not defined for tuples.  In such cases, array
literal syntax should be used.

Finally, there's cleanup of the internal vestiges of 'set' and 'sequence'
types which have been replaced with arrays.