compiler: Refactor the IR compiler to use the functional style and singledispatch.
Currently, the IR compiler is implemented as a set of mixin classes.
There is no real purpose to using the classes other than holding the
context state, however there are several downsides:  it is hard to
move the code around, and implicit dependence on context makes it
harder to reason about method side effects.

This commit converts the IR compiler to a functional style and swaps the
AST visitor for singledispatch.  The context state has been split into
two parts: the global, relatively static Environment, and the dynamic,
stacked Context part.  Functions only relying on the data from the
former part are declared correspondingly.