Add initial support for Views
This adds a way to declare and store EdgeQL views in the schema.
Semantically schema-level views are identical to views declared in the
WITH clause.  However, schema views allow for an easier reuse and can
also be introspected.

   CREATE VIEW MyView := (
   	SELECT Object { foo := 'bar' } FILTER Object.name = 'spam'
   );

   SELECT MyView { foo, __class__: {name} };

   ...

   {
       "id": "...",
       "foo": "bar",
       "__class__": { "name": "Object" }
   }

This commit only adds the necessary DDL and compiler bits.  Declarative
schema and proper introspection are to be handled in a separate commit.