Streamline handling of path context in the compiler
Current code has a lot of hacks around handling of elements in
tuples, which are not currently handled as proper paths.

This commit consolidates the output-pull pipeline so that all
references, including tuple elements, are handled as path vars.
Various relation generation bits no longer populate the output
dictionary and the result target list, and instead rely on
"get_path_output" to do the right thing.

Another important aspect of this rework is the change to the semantics
of atomic path identity.  Atomic sets are no longer made distinct
by value.