Refactor constraints compilation.
Now we inline constraints parameters right into the constraint
expression, instead of guessing their types.  This fixes type
inference of constraints expressions and allows to use
functions in them.

We also no longer use constraint arguments to define the
subject expression; it has to be defined explicitly now.

This allowed to fix the 'std::enum' constraint to use the new
'std::array_contains' function, instead of the deprecated
'IN' operator for arrays.  This unblocks #8.