Implement `IN` as a proper set operator, remove support for arrays
Current `IN` operator tries to work both as a set membership operator,
_and_ an array membership operator, which is bad, as it introduces
ambiguity in cases where there are sets of _arrays_, e.g.:

    WITH Foo := [1, 2] SELECT [1, 2] IN Foo

This ambiguity is absent in SQL (where IN has this duality also), since
the set membership form requires an explicit subquery.

Thus, implement `IN` as a set membership operator, and nothing else.

Fixes: #8.