schema: Remove `\` as line continuation.
Schema was using `\` as line continuation *only* in `:=` expressions.
The problem with that is that actually `\` is not part of EdgeQL grammar
and it is not allowed anywhere else in the schema grammar. So it's
rather an odd exception. The main purpose of that exception seems to be
in allowing to write extra long one-line strings on multiple lines.