Implement CLI interface for running tests
Now that we have a custom test runner, let's take a step further and
add a customized CLI interface for test runs.

This adds a new top-level CLI entry point: et.  It is designed to
be an entry point to all EdgeDB development-related tools.  The new
test CLI is, thus, invoked as follows:

    et test [options] [test paths...]

The result renderer is pretty rudimentary, but it matches the
capabilities of the standard unittest.TextTestRunner.