schema: Rework migration command registration using metaclass kwargs
Currently, migration command implementations are registered using
a variety of class attributes, _and_ an explicit DeltaDriver instance on
the subject schema metaclass.  This is unnecessarily verbose and
promotes poor source code layout in schema implementation modules, so
clean this up.

Now, migration commands use metaclass kwargs to declare their schema
metaclass and the context class.  This is enough to establish the
schema-metaclass -> migration commands link, so drop the DeltaDriver
contraption altogether.

The size of the diff is disproportinately large due to the moving
around of the code in schema modules, as the migration code has
been moved to the end of each module.