Major rewrite of the EdgeQL compiler for performance and correctness
As evident from the edgeql_scope* tests, the present compiler has
significant issues with handling of non-trivial path scope
configurations, for example in the following query:

    WITH A := {1, 2} SELECT (A, test::User{name, a := A});

the value of `User.a` would be computed incorrectly as `[1, 2]`
regardless of the actual value of the first tuple element.

These, and other, issues stem from the ad-hoc scope tracking mechanism
via `ctx.path_scope_refs` and friends, with rather arbitrary scope
binding rules.

This rewrite started with the final formalization of the EdgeQL
mathematical model, which now builds upon a single axiomatic definition
of how a function applies to its input depending on the declaration of
the input variables.  There are three possible kinds of inputs:
regular element-wise inputs, `SET OF` inputs, and `OPTIONAL` inputs,
such that:

    F(SET OF A, SET OF A): { F({A}, {A}) }

and

    F(A, SET OF A): { F({a}, {a}) | ∀ a ∈ A }

and

    F(A): F({}) ≡ {}

and

    F(OPTIONAL A): F({}) ≢ {}

The `SET OF` qualifier signifies that the function is called once with
the entire set as input, as opposed to element-wise invocation, but
_only_ if the same set was _not_ used as a non-`SET OF` input in the same
expression.  The `OPTIONAL` qualifier signifies that the function will
be called with an empty input and may return a non-empty result, whereas
functions with regular element-wise inputs always map to empty set.

With these definitions it is possible to derive all of EdgeQL operators
and clauses.

The new compiler implementation derives an unambiguous path scope tree
for every expression during the EdgeQL-to-IR phase.  The scope tree is
then used by the SQL backend to produce an equivalent SQL expression.

The new implementation no longer uses CTEs to represent intermediate
value sets, and instead relies on `LATERAL` and subqueries to produce
a query structure that is much closer to optimal than previously.
For example, a path expression will now use a semi-join if link source
is not visible anywhere else in the same expression.  This is much more
efficient than using `DISTINCT`.

The new compiler removes the need for the separate SQL optimization
phase, so the optimizer is now obsolete and will be removed in a
subsequent commit.

This commit xfails several tests that use inline views and several
`GROUP` tests.  These were not deemed worth the effort, as both
the views and `GROUP` are due to be reimplemented.