EdgeQL syntax update.
Update GROUP and FOR statement syntax to disambiguate what is being
computed.

Remove SINGLETON syntax in favor of specifying CARDINALITY in the WITH
clause.

Add SET OF and OPTIONAL qualifiers to function signature declaration.

Make FILTER and ORDER BY clauses available for each argument of an
aggregate function call.

Update pygmets grammar for EdgeQL and EdgeSchema.