EdgeQL compiler rewrite part II
This part of the compiler rewrite is largely concentrated around the
generalization of views.  Now, any "shape" expression is treated as a
view declaration.  For example:  User { foo := '1' } is a declaration
of a view that describes a set of users with a computed link "foo".

The standalone View schema class has been removed, and now views are
treated as simple subclasses with benefits to type inference and
arbitrary nesting of views.

More importantly, views declared on the query level, i. e. in the
`WITH` block, are always evaluated in the use-site context with
corresponding changes to the scope rules.  Essentially, such views are
now very close in behaviour to lexical macros.  Due to this change, the
`DETACHED` keyword has been reintroduced and made into an operator that
makes the compiler evaluate its operand in a schema-level scope,
ignoring everything else in the query.

Another meaning of a shape expression is to describe the output
serialization for an object.  This is now achieved by the special
handling of views declared in a SELECT statement output clause.

The handling of path scopes has been reworked significantly to handle
nested views.

The SELECT statement is now a scope fence, like an aggregate function,
and has, essentially became a generic "set expression", which is a more
intuitive behaviour as opposed to SELECT being completely transparent.
Likewise UNION is now a SET OF for both operands, and IF is a SET OF for
its leaf operands.

There a still some rough edges and known issues on the IR compiler side,
but this is sufficiently stable now to be committed.