Normalize treatment of scope in query-level views
This makes all views (any query expression to the right of `:=`), to be
evaluated in a semi-detached scope, where only symbols in the outer
scope are taken into account.  The use of the symbol denoting _any_ view
is now indistinguishable from a reference to a schema-level set, and
symbols used in the view expression are guaranteed not to interfere:

    WITH U1 := User, U2 := User SELECT (U1, U2) -> cartesian product

The `DETACHED` operator makes the compiler completely ignores all current
scope. Only schema-level symbols are recognized in the `DETACHED`
expression, and it is equivalent to declaring a schema-level view.

Additionally, all implicit `SELECTs` are now treated like explicit
`SELECT` statements from the scope standpoint.