graphql: Add a special helper schema module for GraphQL.
GraphQL makes use of structures that do not have an equivalent in
EdgeQL. Specifically every GraphQL query response is packaged into a
"Query" object with some fields. Also the type system is broken down
into different components so that some special GraphQL placeholder types
are necessary to bridge this gap.