edgeql: Change shape and links syntax.
Link names no longer can be fully qualified. Only short names allowed.

In shapes polymorphic links syntax changes from `Concept::link` to
`[IS Concept.link]`. This is more consistent with similar syntax in
paths.

In shapes backwards (and forward) link specification is no longer
allowed. If there's a need to follow a backwards link, the canonical way
now is to define a computable for it.