Nomenclature revamp
There is currently a fair amount of nomenclature cruft around what's
called what in EdgeDB schema and it's time to clean that up.

The new naming structure is simple and is built around the notion of
*type*.  Every value in EdgeDB has a type, and there are three main
categories of types: scalar types (currently "atoms"),
object types (currently "concepts") and container types
("array", "map" and "tuple").  The "__class__" link is now "__type__",
and all uses of "class" in the schema nomenclature are replaced with
"(schema) object".

I decided to be thorough and rename most source code uses of the old
names as well.  Module names have been intentionally left alone for
a clean rename later.