Require ABSTRACT keyword for attributes, constraints and pointers
The schema and DDL declarations of attributes, constraints, links and
link properties do not create concrete schema objects.  Instead, the
purpose of these declarations is to establish an abstract base to use in
concrete object declarations in parent objects.  Hence, we now require the
"abstract" keyword to precede toplevel declarations to explicitly
indicate that such schema objects are abstract.